Move lyric preview into Graphics namespace.
