+++ osu.Game.Rulesets.Karaoke.Tests/Ranking/TestSceneSaitenResultGraph.cs
@@ -5,7 +5,9 @@ using NUnit.Framework;
 using osu.Framework.Extensions.Color4Extensions;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Shapes;
+using osu.Game.Beatmaps;
 using osu.Game.Rulesets.Karaoke.Statistics;
+using osu.Game.Rulesets.Karaoke.Tests.Beatmaps;
 using osu.Game.Scoring;
 using osu.Game.Tests.Visual;
 using osuTK;
@@ -17,10 +19,12 @@ namespace osu.Game.Rulesets.Karaoke.Tests.Ranking
         [Test]
         public void TestManyDistributedEvents()
         {
-            createTest(new ScoreInfo());
+            var ruleset = new KaraokeRuleset().RulesetInfo;
+            var beatmap = new TestKaraokeBeatmap(ruleset);
+            createTest(new ScoreInfo(), beatmap);
         }
 
-        private void createTest(ScoreInfo score) => AddStep("create test", () =>
+        private void createTest(ScoreInfo score, IBeatmap beatmap) => AddStep("create test", () =>
         {
             Children = new Drawable[]
             {
@@ -29,7 +33,7 @@ namespace osu.Game.Rulesets.Karaoke.Tests.Ranking
                     RelativeSizeAxes = Axes.Both,
                     Colour = Color4Extensions.FromHex("#333")
                 },
-                new SaitenResultGraph(score)
+                new SaitenResultGraph(score, beatmap)
                 {
                     Anchor = Anchor.Centre,
                     Origin = Anchor.Centre,
