+++ osu.Game.Rulesets.Karaoke/Graphics/LyricPreview.cs
@@ -1,0 +1,141 @@
+﻿// Copyright (c) andy840119 <andy840119@gmail.com>. Licensed under the GPL Licence.
+// See the LICENCE file in the repository root for full licence text.
+
+using System.Collections.Generic;
+using System.Linq;
+using osu.Framework.Allocation;
+using osu.Framework.Bindables;
+using osu.Framework.Extensions.IEnumerableExtensions;
+using osu.Framework.Graphics;
+using osu.Framework.Graphics.Containers;
+using osu.Framework.Graphics.Shapes;
+using osu.Framework.Graphics.Sprites;
+using osu.Game.Graphics;
+using osu.Game.Graphics.Containers;
+using osu.Game.Rulesets.Karaoke.Objects;
+using osu.Game.Rulesets.Karaoke.UI.Components;
+using osuTK;
+using osuTK.Graphics;
+
+namespace osu.Game.Rulesets.Karaoke.Graphics
+{
+    public class LyricPreview : CompositeDrawable
+    {
+        public Bindable<LyricLine> SelectedLyricLine { get; private set; } = new Bindable<LyricLine>();
+
+        private readonly FillFlowContainer<ClickableLyric> lyricTable;
+
+        public LyricPreview(IEnumerable<LyricLine> lyrics)
+        {
+            InternalChild = new OsuScrollContainer
+            {
+                RelativeSizeAxes = Axes.Both,
+                Child = lyricTable = new FillFlowContainer<ClickableLyric>
+                {
+                    AutoSizeAxes = Axes.Y,
+                    RelativeSizeAxes = Axes.X,
+                    Direction = FillDirection.Vertical,
+                    Spacing = new Vector2(15),
+                    Children = lyrics.Select(x => new ClickableLyric(x)
+                    {
+                        Selected = false,
+                        Action = () => triggerLyricLine(x)
+                    }).ToList()
+                }
+            };
+
+            SelectedLyricLine.BindValueChanged(value =>
+            {
+                var oldValue = value.OldValue;
+                if (oldValue != null)
+                    lyricTable.Where(x => x.HitObject == oldValue).ForEach(x => { x.Selected = false; });
+
+                var newValue = value.NewValue;
+                if (newValue != null)
+                    lyricTable.Where(x => x.HitObject == newValue).ForEach(x => { x.Selected = true; });
+            });
+        }
+
+        private void triggerLyricLine(LyricLine lyric)
+        {
+            if (SelectedLyricLine.Value == lyric)
+                SelectedLyricLine.TriggerChange();
+            else
+                SelectedLyricLine.Value = lyric;
+        }
+
+        internal class ClickableLyric : ClickableContainer
+        {
+            private const float fade_duration = 100;
+
+            private Color4 hoverTextColour;
+            private Color4 idolTextColour;
+
+            private readonly Box background;
+            private readonly SpriteIcon icon;
+            private readonly PreviewLyricSpriteText previewLyric;
+
+            public ClickableLyric(LyricLine lyric)
+            {
+                AutoSizeAxes = Axes.Y;
+                RelativeSizeAxes = Axes.X;
+                Masking = true;
+                CornerRadius = 5;
+                Children = new Drawable[]
+                {
+                    background = new Box
+                    {
+                        RelativeSizeAxes = Axes.Both
+                    },
+                    icon = new SpriteIcon
+                    {
+                        Anchor = Anchor.CentreLeft,
+                        Origin = Anchor.CentreLeft,
+                        Size = new Vector2(15),
+                        Icon = FontAwesome.Solid.Play,
+                        Margin = new MarginPadding { Left = 5 }
+                    },
+                    previewLyric = new PreviewLyricSpriteText(lyric)
+                    {
+                        Font = new FontUsage(size: 25),
+                        RubyFont = new FontUsage(size: 10),
+                        RomajiFont = new FontUsage(size: 10),
+                        Margin = new MarginPadding { Left = 25 }
+                    }
+                };
+            }
+
+            private bool selected;
+
+            public bool Selected
+            {
+                get => selected;
+                set
+                {
+                    if (value == selected) return;
+
+                    selected = value;
+
+                    background.FadeTo(Selected ? 1 : 0, fade_duration);
+                    icon.FadeTo(Selected ? 1 : 0, fade_duration);
+                    previewLyric.FadeColour(Selected ? hoverTextColour : idolTextColour, fade_duration);
+                }
+            }
+
+            public LyricLine HitObject => previewLyric.HitObject;
+
+            [BackgroundDependencyLoader]
+            private void load(OsuColour colours)
+            {
+                hoverTextColour = colours.Yellow;
+                idolTextColour = colours.Gray9;
+
+                previewLyric.Colour = idolTextColour;
+                background.Colour = colours.Blue;
+                background.Alpha = 0;
+                icon.Colour = hoverTextColour;
+                icon.Alpha = 0;
+            }
+        }
+    }
+}
